Create patch request

+++ server/routes/tasks.js
@@ -88,7 +88,28 @@ router.put("/:id", async (req, res) => {
   }
 });
 
+router.patch("/:id", async (req, res) => {
+  const task = await Task.findById(req.params.id);
+
+  if (!task) return res.status(404).send("Todo not found");
+
+  try {
+    const updatedTask = await Task.findByIdAndUpdate(req.params.id, {
+      isComplete: !task.isComplete,
+    });
+
+    res.send(updatedTask);
+  } catch (error) {
+    res.status(500).send(error.message);
+    console.log(error.message);
+  }
+});
+
 router.delete("/:id", async (req, res) => {
+  const task = await Task.findById(req.params.id);
+
+  if (!task) return res.status(404).send("Todo not found");
+
   try {
     const deletedTask = await Task.findByIdAndDelete(req.params.id);
